Refactored AddAnActivityScreen component to separate the form logic and UI rendering into a separate new ActivityForm component.
